Revert split into separate files; Fix tests
