--- globals.js
@@ -1,44 +1,0 @@
-const globals = {};
-
-const getGlobal = property => {
-	/* istanbul ignore next */
-	if (typeof self !== 'undefined' && self && property in self) {
-		return self;
-	}
-
-	/* istanbul ignore next */
-	if (typeof window !== 'undefined' && window && property in window) {
-		return window;
-	}
-
-	if (typeof global !== 'undefined' && global && property in global) {
-		return global;
-	}
-
-	/* istanbul ignore next */
-	if (typeof globalThis !== 'undefined' && globalThis) {
-		return globalThis;
-	}
-};
-
-const globalProperties = [
-	'Headers',
-	'Request',
-	'Response',
-	'ReadableStream',
-	'fetch',
-	'AbortController',
-	'FormData'
-];
-
-for (const property of globalProperties) {
-	Object.defineProperty(globals, property, {
-		get() {
-			const globalObject = getGlobal(property);
-			const value = globalObject && globalObject[property];
-			return typeof value === 'function' ? value.bind(globalObject) : value;
-		}
-	});
-}
-
-export default globals;
